finished rules, reverted
