+++ app/src/main/AndroidManifest.xml
@@ -49,7 +49,7 @@
         <activity android:name=".PlayerListActivity" />
         <activity android:name=".MainGameDay" />
         <activity
-            android:name=".RulesActivity"
+            android:name=".Rules.RulesActivity"
             android:exported="true" />
         <activity android:name=".MainGameNight"></activity>
     </application>
