+++ app/src/main/java/com/example/keiji/app/activities/Rules/RulesActivity.java
@@ -1,4 +1,4 @@
-package com.example.keiji.app.activities;
+package com.example.keiji.app.activities.Rules;
 
 import android.os.Bundle;
 import android.support.design.widget.TabLayout;
@@ -6,10 +6,10 @@ import android.support.v4.view.ViewPager;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 
+import com.example.keiji.app.activities.R;
 import com.example.keiji.app.objects.RulesPagerAdapter;
 
 public class RulesActivity extends AppCompatActivity {
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
