--- app/src/main/java/com/example/keiji/app/activities/SheriffRules.java
@@ -1,15 +1,0 @@
-package com.example.keiji.app.activities;
-
-import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-public class SheriffRules extends Fragment {
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_sherif_rules, container, false);
-    }
-}
