+++ app/src/main/res/layout/fragment_sherif_rules.xml
@@ -6,10 +6,14 @@
 
     <TextView
         android:id="@+id/sherif_view"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerInParent="true"
-        android:text="Sheriff"
-        android:textAppearance="?android:attr/textAppearanceLarge"/>
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_alignParentStart="true"
+        android:layout_alignParentTop="true"
+        android:layout_marginTop="0dp"
+        android:maxLines="200"
+        android:scrollbars="vertical"
+        android:text="@string/rules_sheriff"
+        android:textAppearance="?android:attr/textAppearanceLarge" />
 
 </RelativeLayout>
