+++ app/src/main/res/layout/fragment_vilager_rules.xml
@@ -6,10 +6,14 @@
 
     <TextView
         android:id="@+id/villager_view"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerInParent="true"
-        android:text="Villager"
-        android:textAppearance="?android:attr/textAppearanceLarge"/>
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_alignParentStart="true"
+        android:layout_alignParentTop="true"
+        android:layout_marginTop="0dp"
+        android:maxLines="200"
+        android:scrollbars="vertical"
+        android:text="@string/rules_villager"
+        android:textAppearance="?android:attr/textAppearanceLarge" />
 
 </RelativeLayout>
